third_party/blink/web_tests/platform/generic/external/wpt/scroll-animations/css/at-scroll-timeline-paused-animations-expected.txt

This is a testharness.js-based test.
FAIL Test that the scroll animation is paused assert_equals: expected "100px" but got "50px"
FAIL Test that the scroll animation is paused by updating animation-play-state assert_equals: expected "100px" but got "50px"
Harness: the test ran to completion.
